Move deployment utilities to there own package
